+++ src/app/cart/cart.module.ts
@@ -1,17 +1,13 @@
 import { NgModule } from '@angular/core';
+
 import { SharedModule } from '../shared/shared.module';
-import { CartItemComponent } from './components/cart-item/cart-item.component';
-import { CartListComponent } from './components/cart-list/cart-list.component';
+import { CartItemComponent, CartListComponent } from './index';
 
 @NgModule({
     imports: [
         SharedModule
     ],
     declarations: [
-        CartItemComponent,
-        CartListComponent
-    ],
-    exports: [
         CartItemComponent,
         CartListComponent
     ]