+++ src/app/cart/services/cart.service.ts
@@ -2,13 +2,34 @@ import { Injectable } from '@angular/core';
 
 import { ProductModel } from 'src/app/products/models/product.model';
 import { CartItemModel } from '../models/cart-item.model';
+import { LocalStorageService } from 'src/app/shared/services/local-storage.service';
 
 @Injectable({
     providedIn: 'root'
 })
 export class CartService {
 
+    private readonly cartItemsStorageKey = 'cartItems';
     private cartItems: CartItemModel[] = [];
+
+    constructor(private readonly storage: LocalStorageService) {
+        this.cartItems = storage.getItem<Array<CartItemModel>>(this.cartItemsStorageKey) || [];
+    }
+
+    get totalQuantity(): number {
+        return this.cartItems
+            .map((item) => item.quantity)
+            .reduce((prev, next) => prev + next, 0);
+    }
+
+    get isEmpty(): boolean {
+        return this.totalQuantity === 0;
+    }
+
+    get totalSum(): number {
+        return this.cartItems.map(this.getCartItemTotalPrice)
+            .reduce((prev, next) => prev + next, 0);
+    }
 
     getCartItems(): ReadonlyArray<CartItemModel> {
         return this.cartItems;
@@ -18,25 +39,24 @@ export class CartService {
         return cartItem.product.price * cartItem.quantity;
     }
 
-    getCartTotalPrice(): number {
-        return this.cartItems.map(this.getCartItemTotalPrice)
-            .reduce((prev, next) => prev + next, 0);
-    }
-
     addProductToCart(product: ProductModel): void {
         const existingItem = this.searchCartForProductByName(product.name);
         if (existingItem !== undefined) {
             this.increaseQuantityByOne(product);
         } else {
             this.cartItems.push(this.createNewCartItem(product));
+            this.updateCart();
         }
     }
 
     increaseQuantityByOne(product: ProductModel): void {
         const itemToIncrease = this.searchCartForProductByName(product.name);
-        if (itemToIncrease !== undefined) {
-            itemToIncrease.quantity++;
+        if (itemToIncrease === undefined) {
+            return;
         }
+
+        itemToIncrease.quantity++;
+        this.updateCart();
     }
 
     decreaseQuantityByOne(product: ProductModel): void {
@@ -50,11 +70,23 @@ export class CartService {
         } else {
             itemToDecrease.quantity--;
         }
+
+        this.updateCart();
     }
 
     removeProductFromCart(product: ProductModel): void {
         const productToRemove = this.searchCartForProductByName(product.name);
         this.cartItems = this.cartItems.filter(item => item !== productToRemove);
+        this.updateCart();
+    }
+
+    removeAllProducts(): void {
+        this.cartItems = [];
+        this.updateCart();
+    }
+
+    updateCart(): void {
+        this.storage.setItem<Array<CartItemModel>>(this.cartItemsStorageKey, this.cartItems);
     }
 
     private searchCartForProductByName(productName: string): CartItemModel {