+++ src/app/shared/material.module.ts
@@ -12,18 +12,6 @@ import { MatInputModule } from '@angular/material/input';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 
 @NgModule({
-    imports: [
-        MatToolbarModule,
-        MatIconModule,
-        MatGridListModule,
-        MatMenuModule,
-        MatButtonModule,
-        MatSidenavModule,
-        MatCardModule,
-        MatListModule,
-        MatInputModule,
-        MatSnackBarModule
-    ],
     exports: [
         MatToolbarModule,
         MatIconModule,